ON-14765: Allocate EFCT memory with hugetlb library
This patch must be a no-op for the EFCT workflow, except for how we get
a page reference. It used to be find_get_page(), but now it is
pin_user_pages() that is preferable as it migrates the allocated pages
away from the movable zones preserving their mobility, but requires a
userspace address which we get by vm_mmap(), assuming the allocation
occurs in the application context. If this is not the case, we should
revert the hugetlb library to use find_get_page().

It also changes the ef_vi interface by removing the unneeded
in_memfd_off parameter.